fix: Corregir errores de tipos en clases del calendario
 Cambiar opacidades de int a double en CalendarConstants
 Corregir tipo de retorno en CalendarUtils.getTrackColor
 Verificar tipos correctos en CalendarEventLogic
 Sin errores de compilación
 Integración funcionando correctamente

Archivos corregidos:
- calendar_constants.dart: Opacidades como double
- calendar_utils.dart: Verificación de tipo Color
- calendar_event_logic.dart: Tipos correctos

+++ lib/widgets/screens/calendar/calendar_constants.dart
@@ -33,6 +33,6 @@ class CalendarConstants {
   static const int defaultVisibleDays = 7;
   
   // Colores de estado
-  static const int todayHighlightOpacity = 0.1;
-  static const int trackHighlightOpacity = 0.05;
+  static const double todayHighlightOpacity = 0.1;
+  static const double trackHighlightOpacity = 0.05;
 }
